Documentation
The Air Quality Monitoring System is a significant innovation in the realm of
environmental sensing and data visualization. This project leverages the power of an
Arduino UNO microcontroller, an MQ-135 gas sensor, a breadboard, connecting
jumper wires, an LED, a buzzer, a 16x2 LCD monitor, and an array of components
to create a comprehensive solution for monitoring air quality. Air quality has a
profound impact on our health and well-being, making it vital to keep a close eye on
its status. This system not only detects air quality but also provides real-time
feedback, indicating whether the air quality falls within acceptable levels.